Fix non-ortogonal line after re-route and non-interactive toggle
Also moves path hooks to own file as use-interactive module was having to many different functions
